Flask/requirements.txt
flask
python-dotenv
SQLAlchemy==1.4.49
psycopg2-binary
pydantic==1.10.9
